==> movie_al/01.ReadExtractIMDBReviews.ipynb <==
# ReadExtractDataIMDBReviews

Download the IMDB dataset and perform basic preparation.

import os
import re
import urllib
import tarfile
import sqlite3

import pandas as pd

## Download The Dataset

os.makedirs("var", exist_ok=True)

data_url = "https://ai.stanford.edu/~amaas/data/sentiment/aclImdb_v1.tar.gz"
local_file = "var/aclImdb_v1.tar.gz"

if not(os.path.isfile(local_file)):
    urllib.request.urlretrieve(data_url, local_file)

## Dataset Reading Functions

def read_text(tf, name):
    with tf.extractfile(name) as f:
        return f.read().decode('utf-8')

def read_readme(tf):
    return read_text(tf, "aclImdb/README")

def read_comment_links(tf, name):
    comments = []
    with tf.extractfile(name) as f:
        for line in f:
            comments.append(line.decode('utf-8').strip())
    return comments

def read_all_comment_links(tf):
    dfs = []
    names = tf.getnames()
    for ds_type in ['train', 'test']:
        for _class in ['pos', 'neg', 'unsup']:
            fname = f"aclImdb/{ds_type}/urls_{_class}.txt"
            if fname in names:
                links = read_comment_links(tf, fname)
                links_df = pd.DataFrame({'commentLink': links})
                links_df['dataset'] = ds_type
                links_df['class'] = _class
                links_df['datasetclassId'] = range(0, len(links_df))
                dfs.append(links_df)
            
    return pd.concat(dfs)

def get_review_names(tf):
    name_re = re.compile(r'^aclImdb/(test|train)/(pos|neg|unsup)/(\d+)_(\d+).txt$')

    entries = []
    for name in tf.getnames():
        m = name_re.match(name)
        if m:
            ds_type = m.group(1)
            _class = m.group(2)
            _id = int(m.group(3))
            rating = int(m.group(4))
            entries.append({'dataset': ds_type, 'class': _class, 'datasetclassId': _id, 'rating': rating, 'name': name})

    return pd.DataFrame(entries)

def read_reviews(tf):
    reviews = get_review_names(tf)
    reviews['review'] = reviews['name'].apply(lambda name: read_text(tf, name))
    return reviews

## Open Handle to Tarfile

tf = tarfile.open(local_file)

### Readme

print(read_readme(tf))

### Load Comment Links

comment_links = read_all_comment_links(tf)

comment_links['class'].value_counts().to_frame()

comment_links.sample(5)

### Load Reviews

reviews = read_reviews(tf)

reviews.groupby(['dataset', 'class']).size().reset_index()

reviews.sample(5)

### Final Prep and Join Comments Links with Reviews

ttid_re = re.compile(r'title/([^/]+)')
                     
def get_ttid_re(comment_link):
    m = ttid_re.search(comment_link)
    if m:
        return m.group(1)

reviews_with_comments = pd.merge(
    reviews,
    comment_links
)

# Assign a pk
reviews_with_comments['id'] = range(1, len(reviews_with_comments) + 1)

# Extract the title id
reviews_with_comments['titleId'] = reviews_with_comments.commentLink.apply(get_ttid_re)

# Drop some columns and reorde columns
reviews_with_comments = (reviews_with_comments
    [['id', 'titleId', 'dataset', 'class', 'datasetclassId', 'rating', 'review']]
)

reviews_with_comments.sample(5)

## Write Data Out to Sqlite

db = sqlite3.connect("var/reviews.db")

(reviews_with_comments
    .to_sql('reviews', if_exists='replace', index=False, chunksize=5000, con=db)
)

db.commit()

db.execute("create unique index i_reviews_pk on reviews (id)")

db.execute("create index i_reviews_title_id on reviews (titleId)")

db.commit()

db.close()
