==> movie_al/02.ReadExtractIMDBTitleInfo.ipynb <==
# ReadExtractIMDBTitleInfo

import os
import re
import urllib
import tarfile
import sqlite3
import gzip
import csv

import pandas as pd

## Download Data

os.makedirs("var", exist_ok=True)

data_url = "https://datasets.imdbws.com/title.basics.tsv.gz"
local_file = "var/title.basics.tsv.gz"

if not(os.path.isfile(local_file)):
    urllib.request.urlretrieve(data_url, local_file)

## Load Data

title_info = pd.read_csv(local_file, sep='\t', encoding='utf-8', na_values=['\\N'], quoting=csv.QUOTE_NONE)

title_info.sample(5)

title_info.isAdult.value_counts().to_frame()

title_info.titleType.value_counts().to_frame()

title_info.startYear.value_counts().to_frame()

title_info = title_info.rename(columns={'tconst': 'titleId'})

title_info

## Save to DB

db = sqlite3.connect("var/reviews.db")

title_info.to_sql('titles', if_exists='replace', index=False, chunksize=5000, con=db)

db.commit()

db.execute("create unique index i_titles_pk on titles (titleId)")
db.commit()

db.close()
