==> movie_al/03.VectorizeReviews.ipynb <==
# Create Vector Representations for Review Text

import os
import re
import urllib
import tarfile
import sqlite3
import spacy
import logging
import sys
import json

import pandas as pd

import gensim
from gensim.models.doc2vec import TaggedDocument, Doc2Vec

import spacy

## Setup Logging

# Add logging at the info level
if "_LOG_SETUP" not in globals():
    root_logger = logging.getLogger()
    root_logger.setLevel(logging.INFO)

    log_handler = logging.StreamHandler(sys.stdout)
    log_handler.setLevel(logging.INFO)
    log_formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
    log_handler.setFormatter(log_formatter)
    root_logger.addHandler(log_handler)
    _LOG_SETUP = True

## Load Data

db = sqlite3.connect("var/reviews.db")

reviews = pd.read_sql("select * from reviews", db)

reviews.sample(5)

## Get Words

emb_tag_re = re.compile(r'<(br)\s*/?\s*>')

def replace_tags(text):
    return emb_tag_re.sub('\n', text)

word_split_re = re.compile(r'[,\.\s\?\!":;\(\)\[\]]')

def get_words_simple(text):
    words = []
    for w in word_split_re.split(replace_tags(text)):
        w = w.lower().strip()
        if len(w) > 0:
            words.append(w)
    return words

reviews['words'] = reviews.review.apply(get_words_simple)

reviews.sample(5)

## Train doc2vec

docs = []
for _, row in reviews.iterrows():
    docs.append(TaggedDocument(row.words, [f"{row.id}"]))

%%time
doc_model = Doc2Vec(docs, vector_size=50, epochs=10)

os.makedirs("var/models", exist_ok=True)
doc_model.save("var/models/review_model.d2v")

## Get Embeddings

reviews['review_embedding'] = reviews.words.apply(lambda words: doc_model.infer_vector(words))

## Save embeddings by id to SQL

review_saves = reviews[['id', 'review_embedding']].copy()

#review_saves['words'] = review_saves.words.apply(json.dumps)

review_saves['review_embedding'] = review_saves.review_embedding.apply(lambda emb: json.dumps([float(e) for e in emb]))

review_saves.sample(5)

db = sqlite3.connect("var/reviews.db")

review_saves.to_sql('review_embeddings', if_exists='replace', index=False, chunksize=5000, con=db)

db.commit()

db.execute("create unique index i_review_embeddings_pk on review_embeddings (id)")
db.commit()

db.close()
