==> movie_al/04.ALWantToWatchTraining.ipynb <==
# Train Model to Find Movies I Want to Watch

https://github.com/modAL-python/modAL

import os
import re
import urllib
import tarfile
import sqlite3
import spacy
import logging
import sys
import json
import pickle

import pandas as pd
import numpy as np

import modAL
from modAL.models import ActiveLearner
from modAL.uncertainty import uncertainty_sampling, entropy_sampling
from modAL.density import information_density

from sklearn.ensemble import RandomForestClassifier

from IPython.display import display, clear_output, HTML, JSON

## Config

reviews_db = "var/reviews.db"
label_table = "watch_labels"

## Load Data

db = sqlite3.connect(reviews_db)

pd.read_sql("select * from sqlite_master", db)

reviews = pd.read_sql("""
    select
        r.*,
        e.review_embedding
    from reviews r
    left join review_embeddings e on (e.id = r.id)
""", db)

reviews['review_embedding'] = reviews.review_embedding.apply(lambda et: np.array(json.loads(et)))

reviews.sample(5)

titles = pd.read_sql("""
    select *
    from titles t
    where
        t.titleId in (
            select titleId
            from reviews
        )
""", db)

titles.sample(5)

reviews = pd.merge(
    reviews,
    titles[['titleId', 'titleType', 'primaryTitle', 'genres', 'startYear']],
    on=['titleId'],
    how = 'left'
)

reviews.titleType.value_counts().to_frame()

# I want to limit this to movies
reviews = reviews[reviews.titleType.isin({'movie', 'tvMovie'})]

len(reviews)

reviews = reviews.set_index('id')

## Active Learning Code Modules

### Label Management

Responsible for updating and retrieving labels.

class LabelManagement:
    
    def __init__(self, db, label_table, reviews):
        self.db = db
        self.label_table = label_table
        self.reviews = reviews
        self.initialize()
        
    def initialize(self):
        self.initialize_label_table()
        self.load_labels()
        self.update_pool_and_training()
        
    def clear_labels(self):
        self.db.execute(f"delete from {self.label_table}")
        self.db.commit()
        self.update_pool_and_training()
        
    def initialize_label_table(self):
        lbl_table_count = pd.read_sql(
            "select count(*) as Count from sqlite_master where type = 'table' and name = ?",
            self.db,
            params=[self.label_table]
        ).iloc[0].Count

        if lbl_table_count < 1:
            self.db.execute(f"""
                create table {self.label_table} (
                    id varchar(100) not null primary key,
                    wanted int not null
                )
            """)
            self.db.commit()
         
    def save_labels(self):
        self.labels.to_sql(label_table, db, if_exists = 'replace')
        self.db.execute(f"""
            create unique index i_{self.label_table}_pk on {self.label_table} (id)
        """)
        self.db.commit()
        
    def load_labels(self):
        self.labels = pd.read_sql(f"select * from {self.label_table}", self.db)
        self.labels = self.labels.set_index('id')
        return self.labels
        
    def update_pool_and_training(self):
        labelled_ids = set(self.labels.index.values)
        self.pool = self.reviews[~self.reviews.index.isin(labelled_ids)]
        self.train = self.reviews[self.reviews.index.isin(labelled_ids)]
        
        self.x_pool = np.vstack(self.pool.review_embedding)
        self.x_pool_ids = self.pool.index.values
        
        if len(self.train) > 0:
            self.x_train = np.vstack(self.train.review_embedding)
            self.x_train_ids = self.train.index.values
            self.y_train = self.labels.loc[self.x_train_ids].wanted.values
        else:
            self.x_train = None
            self.x_train_ids = None
            self.y_train = None
        
    def add_labels(self, idxs, results):
        self.labels = pd.concat([self.labels, pd.DataFrame({'id': self.x_pool_ids[idxs], 'wanted': results}).set_index('id')])
        self.update_pool_and_training()
        
lbl_manager = LabelManagement(db, label_table, reviews)

lbl_manager.labels

# Run the line below to clear out previous labels

#lbl_manager.clear_labels()

### LabelUI

Responsible for handling basic input interactions with the user to obtain labels.

class LabelUI:
    def __init__(self, lbl_manager):
        self.lbl_manager = lbl_manager
        
    def query_wanted(self, pool_idx):
        while True:
            #clear_output()
            r_id = self.lbl_manager.x_pool_ids[pool_idx]
            r = self.lbl_manager.reviews.loc[r_id]
            print(r.review)
            print("\n")
            print("Do you want to watch this? [y/n]")

            ans = input().lower().strip()
            if ans == 'y':
                return 1
            elif ans == 'n':
                return 0

    def query_batch(self, idxs):
        results = [self.query_wanted(idx) for idx in idxs]
        x = self.lbl_manager.x_pool[idxs]
        self.lbl_manager.add_labels(idxs, results)
        return (x, idxs, results)

    def query_random(self, n):
        idxs = pd.Series(range(0, len(self.lbl_manager.x_pool_ids))).sample(n).values
        return self.query_batch(idxs)

lbl_ui = LabelUI(lbl_manager)

### ActiveLearnerDriver

Responsible for putting the label management and UI together to drive the overall experience of labelling.

class ActiveLearnerDriver:
    
    def __init__(self, learner, lbl_ui):
        self.lbl_ui = lbl_ui
        self.learner = learner
        
    def query_random(self, n):
        lbl_ui.query_random(n)
        lbl_manager = self.lbl_ui.lbl_manager
        learner.teach(lbl_manager.x_train, lbl_manager.y_train.astype(bool))
        lbl_manager.save_labels()
        
    def query_n(self, n):
        lbl_manager = self.lbl_ui.lbl_manager
        for round in range(0, n):
            query_idx, query_inst = learner.query(lbl_manager.x_pool)
            print(f"idx: {query_idx}")
            x, ids, results = lbl_ui.query_batch(query_idx)
            learner.teach(x, np.array(results).astype(bool), only_new=False)
        lbl_manager.save_labels()

## Active Learner Setup

Create and setup the active learner.

def sampling_factory(al, x):
    return entropy_sampling(al, x, n_instances=1, random_tie_break=True)
    
model = RandomForestClassifier()

learner = ActiveLearner(
    estimator = model,
    query_strategy = sampling_factory
)

al_driver = ActiveLearnerDriver(learner, lbl_ui)

## Get Labels
The cell below well figure out which label to get next, prompt you for a label, and save the results.

al_driver.query_n(1)

# As you label you can see the feature importances change
model.feature_importances_

# See how many labels we have in the db
pd.read_sql(f"select count(*) from {label_table}", db)

### Get Predictions

Using the model, get predictions about a review being compelling.

review_p = model.predict_proba(np.vstack(lbl_manager.reviews.review_embedding)).T
reviews_wp = lbl_manager.reviews.copy()
reviews_wp['bad_p'] = review_p[0]
reviews_wp['good_p'] = review_p[1]

reviews_wp = pd.merge(
    reviews_wp,
    lbl_manager.labels,
    how = 'left',
    left_index = True,
    right_index = True
)

reviews_wp = reviews_wp.sort_values(by='good_p', ascending=False)

### Show top 20 movies to watch which weren't explicitlyl labelled.

reviews_wp[reviews_wp.wanted.isnull()][['primaryTitle', 'rating', 'genres', 'startYear', 'good_p', 'review']].head(20)

# Show top 20 whether explictlyl labelled or not
reviews_wp[['primaryTitle', 'rating', 'genres', 'startYear', 'good_p', 'review', 'wanted']].head(20)

# Show worst 10
reviews_wp[['primaryTitle', 'rating', 'genres', 'startYear', 'good_p', 'review', 'wanted']].tail(10)

# Wanted to see how genres relate to chance of wanting to watch.

genres_good_p = (reviews_wp
    .groupby('genres')
    .agg({'good_p': ['mean', len]})
    .reset_index()
    .sort_values(by=[('good_p', 'mean')], ascending=False))

genres_good_p[genres_good_p[('good_p', 'len')] >= 20].head(20)

# Save Results

# Save results to csv.

movies_to_watch = reviews_wp[['titleId', 'primaryTitle', 'startYear', 'genres', 'review', 'good_p', 'rating']]
movies_to_watch.to_csv("var/movies_wanted.csv")

# Save Model

with open("var/models/wanted_review.pkl", "wb") as fh:
    pickle.dump(model, fh)
